+++ lib/src/tools/format_tool.dart
@@ -13,7 +13,6 @@ import 'package:yaml/yaml.dart';
 import '../dart_dev_tool.dart';
 import '../utils/arg_results_utils.dart';
 import '../utils/assert_no_positional_args_nor_args_after_separator.dart';
-import '../utils/dart_semver_version.dart';
 import '../utils/executables.dart' as exe;
 import '../utils/logging.dart';
 import '../utils/organize_directives/organize_directives_in_paths.dart';
@@ -64,14 +63,13 @@ class FormatTool extends DevTool {
   List<Glob>? exclude;
 
   /// The formatter to run, one of:
-  /// - `dartfmt` (provided by the SDK)
+  /// - `dart format` (provided by the SDK)
   /// - `dart run dart_style:format` (provided by the `dart_style` package)
-  /// - `dart format` (added in Dart SDK 2.10.0)
-  Formatter formatter = Formatter.dartfmt;
+  Formatter formatter = Formatter.dartFormat;
 
   /// The args to pass to the formatter process run by this command.
   ///
-  /// Run `dartfmt -h -v` or `dart format -h -v` to see all available args.
+  /// Run `dart format -h -v` to see all available args.
   List<String>? formatterArgs;
 
   /// The language version to pass to formatters that support
@@ -116,17 +114,19 @@ class FormatTool extends DevTool {
     ..addOption(
       'formatter-args',
       help:
-          'Args to pass to the "dartfmt" or "dart format" process.\n'
-          'Run "dartfmt -h -v" or "dart format -h -v" to see all available options.',
+          'Args to pass to the "dart format" process.\n'
+          'Run "dart format -h -v" to see all available options.',
     );
 
   @override
-  String? description = 'Format dart files in this package.';
+  String? get description =>
+      super.description ?? 'Format dart files in this package.';
 
   @override
   FutureOr<int?> run([DevToolExecutionContext? context]) async {
     context ??= DevToolExecutionContext();
-    if (formatter == Formatter.dartfmt && !dartVersionHasDartfmt) {
+    // ignore: deprecated_member_use_from_same_package
+    if (formatter == Formatter.dartfmt) {
       formatter = Formatter.dartFormat;
     }
     final formatExecution = buildExecution(
@@ -200,7 +200,10 @@ class FormatTool extends DevTool {
     bool? expandCwd,
     bool? followLinks,
     String? root,
-    @deprecated bool? collapseDirectories,
+    @Deprecated(
+      'Ignored; argv length is handled by parallel formatter invocations',
+    )
+    bool? collapseDirectories,
   }) {
     if (collapseDirectories != null) {
       _log.warning(
@@ -253,9 +256,12 @@ class FormatTool extends DevTool {
 class FormatterInputs {
   FormatterInputs(
     this.includedFiles, {
-    @deprecated this.excludedFiles,
-    @deprecated this.hiddenDirectories,
-    @deprecated this.skippedLinks,
+    @Deprecated('Always empty after dart_dev 4.x; see Workiva/dart_dev#424')
+    this.excludedFiles,
+    @Deprecated('Always empty after dart_dev 4.x; see Workiva/dart_dev#424')
+    this.hiddenDirectories,
+    @Deprecated('Always empty after dart_dev 4.x; see Workiva/dart_dev#424')
+    this.skippedLinks,
   });
 
   final Set<String> includedFiles;
@@ -263,13 +269,13 @@ class FormatterInputs {
   // These fields are deprecated and are likely to be empty, due to
   // performance optimizations made in
   // https://github.com/Workiva/dart_dev/pull/424
-  @deprecated
+  @Deprecated('Always empty after dart_dev 4.x; see Workiva/dart_dev#424')
   final Set<String>? excludedFiles;
 
-  @deprecated
+  @Deprecated('Always empty after dart_dev 4.x; see Workiva/dart_dev#424')
   final Set<String>? hiddenDirectories;
 
-  @deprecated
+  @Deprecated('Always empty after dart_dev 4.x; see Workiva/dart_dev#424')
   final Set<String>? skippedLinks;
 }
 
@@ -315,21 +321,24 @@ class DirectiveOrganization {
 
 /// Modes supported by the dart formatter.
 enum FormatMode {
-  // dartfmt -n --set-exit-if-changed
+  // dart format --set-exit-if-changed
   check,
-  // dartfmt -n
+  // dart format -o none
   dryRun,
-  // dartfmt -w
+  // dart format (writes files)
   overwrite,
 }
 
 /// Available dart formatters.
 enum Formatter {
-  // The formatter provided via the Dart SDK.
+  /// The `dartfmt` SDK command, which was removed in Dart 2.15.
+  ///
+  /// Treated as [dartFormat].
+  @Deprecated('dartfmt was removed from the SDK; use Formatter.dartFormat')
   dartfmt,
   // The formatter provided via the `dart_style` package.
   dartStyle,
-  // The formatter provided via the Dart 2.10 SDK
+  // The formatter provided via the Dart SDK (`dart format`).
   dartFormat,
 }
 
@@ -360,7 +369,7 @@ Iterable<String> buildArgs(
   final args = <String>[
     ...executableArgs,
 
-    // Combine all args that should be passed through to the dartfmt in this
+    // Combine all args that should be passed through to the formatter in this
     // order:
     // 1. Mode flag(s), if configured
     if (mode == FormatMode.check) ...['-n', '--set-exit-if-changed'],
@@ -469,12 +478,18 @@ FormatExecution buildExecution(
       allowRest: useRestForInputs,
       commandName: context.commandName,
       usageFooter:
-          'Arguments can be passed to the "dartfmt" or "dart format" process via the '
+          'Arguments can be passed to the "dart format" process via the '
           '--formatter-args option.',
     );
     mode = validateAndParseMode(argResults, context.usageException);
   }
   mode ??= defaultMode;
+
+  formatter ??= Formatter.dartFormat;
+  // ignore: deprecated_member_use_from_same_package
+  if (formatter == Formatter.dartfmt) {
+    formatter = Formatter.dartFormat;
+  }
 
   if (formatter == Formatter.dartStyle &&
       !packageIsImmediateDependency('dart_style', path: path)) {
@@ -482,10 +497,10 @@ FormatExecution buildExecution(
       red.wrap('Cannot run "dart_style:format".\n')! +
           yellow.wrap(
             'You must either have a dependency on "dart_style" in '
-            'pubspec.yaml or configure the format tool to use "dartfmt" '
+            'pubspec.yaml or configure the format tool to use "dart format" '
             'instead.\n'
             'Either add "dart_style" to your pubspec.yaml or configure the '
-            'format tool to use "dartfmt" instead.',
+            'format tool to use "dart format" instead.',
           )!,
     );
     return FormatExecution.exitEarly(ExitCode.config.code);
@@ -565,18 +580,17 @@ FormatExecution buildExecution(
 /// Returns a representation of the process that will be run by [FormatTool]
 /// based on the given [formatter].
 ///
-/// - [Formatter.dartfmt] -> `dartfmt`
-/// - [Formatter.dartFormat] -> `dart format`
+/// - [Formatter.dartfmt] / [Formatter.dartFormat] -> `dart format`
 /// - [Formatter.dartStyle] -> `dart run dart_style:format`
 ProcessDeclaration buildFormatProcess([Formatter? formatter]) {
   switch (formatter) {
     case Formatter.dartStyle:
       return ProcessDeclaration(exe.dart, ['run', 'dart_style:format']);
-    case Formatter.dartFormat:
-      return ProcessDeclaration(exe.dart, ['format']);
+    // ignore: deprecated_member_use_from_same_package
     case Formatter.dartfmt:
+    case Formatter.dartFormat:
     default:
-      return ProcessDeclaration(exe.dartfmt, []);
+      return ProcessDeclaration(exe.dart, ['format']);
   }
 }
 
@@ -613,7 +627,7 @@ String? _formatterLanguageVersion(
   if (formatter == Formatter.dartStyle && !dartStyleSupportsWriteArg) {
     return configuredLanguageVersion ?? 'latest';
   }
-  if (formatter == Formatter.dartFormat && dartSemverVersion.major >= 3) {
+  if (formatter == Formatter.dartFormat) {
     return configuredLanguageVersion ?? 'latest';
   }
   return null;