+++ lib/src/utils/executables.dart
@@ -1,7 +1,1 @@
-import 'dart:io';
-
 final dart = 'dart';
-
-final dartanalyzer = Platform.isWindows ? 'dartanalyzer.bat' : 'dartanalyzer';
-
-final dartfmt = Platform.isWindows ? 'dartfmt.bat' : 'dartfmt';