+++ lib/src/utils/parse_imports.dart
@@ -1,10 +1,8 @@
-import 'package:collection/collection.dart';
-
 /// Return the contents of the enquoted portion of the import statements in the
 /// file. Not 100% accurate, since we use regular expressions instead of the
 /// Dart AST to extract the imports.
 Iterable<String> parseImports(String fileContents) =>
-    _importRegex.allMatches(fileContents).map((m) => m.group(1)).whereNotNull();
+    _importRegex.allMatches(fileContents).map((m) => m.group(1)).nonNulls;
 
 final _importRegex = RegExp(
   r'''^import ['"]([^'"]+)['"];?$''',
@@ -14,7 +12,7 @@ final _importRegex = RegExp(
 /// Return a set of package names given a list of imports.
 Set<String> computePackageNamesFromImports(Iterable<String> imports) => imports
     .map((i) => _packageNameFromImportRegex.matchAsPrefix(i)?.group(1))
-    .whereNotNull()
+    .nonNulls
     .toSet();
 
 final _packageNameFromImportRegex = RegExp(r'package:([^/]+)/.+');