+++ test/functional/fixtures/null_safety/opted_in_custom_config_version_comment/tool/dart_dev/config.dart
@@ -1,6 +1,0 @@
-// @dart=2.9
-import 'package:dart_dev/dart_dev.dart';
-
-final config = {
-  ...coreConfig,
-};