+++ test/tools/function_tool_test.dart
@@ -1,4 +1,5 @@
 @TestOn('vm')
+library;
 import 'package:args/args.dart';
 import 'package:args/command_runner.dart';
 import 'package:dart_dev/dart_dev.dart';