+++ test/tools/webdev_serve_tool_test.dart
@@ -1,9 +1,9 @@
 @TestOn('vm')
+library;
 import 'package:args/args.dart';
 import 'package:args/command_runner.dart';
 import 'package:dart_dev/src/dart_dev_tool.dart';
 import 'package:dart_dev/src/tools/webdev_serve_tool.dart';
-import 'package:dart_dev/src/utils/dart_semver_version.dart';
 import 'package:dart_dev/src/utils/executables.dart' as exe;
 import 'package:io/ansi.dart';
 import 'package:io/io.dart';
@@ -169,7 +169,7 @@ void main() {
       pubCacheWithWebdev = TempPubCache();
       globalActivate(
         'webdev',
-        '^${dartSemverVersion.major}.0.0',
+        webdevVersionConstraint,
         environment: pubCacheWithWebdev.envOverride,
       );
 
@@ -226,7 +226,7 @@ void main() {
                 allOf(
                   contains('webdev serve could not run'),
                   contains(
-                    'dart pub global activate webdev ^${dartSemverVersion.major}.0.0',
+                    "dart pub global activate webdev '$webdevVersionConstraint'",
                   ),
                 ),
               ),